chore: move the routes directory from /routes to src/routes and rename the files to reflect their purpose as routes (e.g. auth.routes.js)
