+++ src/routes/auth.routes.js
